+++ protostar/protostar_toml/sections/__init__.py
@@ -1,0 +1,3 @@
+from protostar.protostar_toml.sections.protostar_config import ProtostarConfig
+from protostar.protostar_toml.sections.protostar_contracts import ProtostarContracts
+from protostar.protostar_toml.sections.protostar_project import ProtostarProject
