+++ protostar/protostar_toml/sections/protostar_project.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+from pathlib import Path
+from typing import Any, Dict
+
+
+@dataclass
+class ProtostarProject:
+    libs_path: Path
+
+    @classmethod
+    def from_dict(cls, raw_dict: Dict[str, Any]) -> "ProtostarProject":
+        return cls(libs_path=Path(raw_dict["libs_path"]))
